requirements.txt
pandas
numpy
scikit-learn
scipy
matplotlib
seaborn
plotly
kaleido
wordcloud
mlxtend
nltk
Flask
Flask-WTF
flask_sqlalchemy
SQLAlchemy
mysql-connector-python
PyMySQL
python-dotenv
joblib
requests
pyarrow
tabulate
celery==5.3.4
redis==5.0.1
schedule==1.2.0
Flask-Limiter==3.5.0
